# app/definition/swagger.yaml
swagger: "2.0"
info:
  version: "1.0"
  title: OnJobs
  description: API de servicios de OnJobs
# during dev, should point to your local machine
host: localhost:3000
...
basePath: /onjobs/v1
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /users/:
    get:
      tags:
        - Usuario
      description: Listar todos los usuarios
      # used as the method name of the controller
      operationId: findAllUsers
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Usuario
      description: Registrar usuario
      # used as the method name of the controller
      operationId: addUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /users/:id/:
    get:
      tags:
        - Usuario
      description: Obtener usuario (ID de usuario)
      # used as the method name of the controller
      operationId: findById
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Usuario
      description: Actualizar usuario (ID de usuario)
      # used as the method name of the controller
      operationId: updateUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Usuario
      description: Eliminar usuario (ID de usuario)
      # used as the method name of the controller
      operationId: deleteUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /cv/experience/user/:id/:
    get:
      tags:
        - Curriculum Vitae
      description: Obtener todas las experiencias (ID de usuario)
      # used as the method name of the controller
      operationId: findExpByIdUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      tags:
        - Curriculum Vitae
      description: Registrar experiencia de usuario (ID de usuario)
      # used as the method name of the controller
      operationId: addExpUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserExp"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /cv/experience/:id/:
    get:
      tags:
        - Curriculum Vitae
      description: Obtener experiencia (ID de experiencia)
      # used as the method name of the controller
      operationId: findExpById
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    put:
      tags:
        - Curriculum Vitae
      description: Actualizar experiencia de usuario (ID de experiencia)
      # used as the method name of the controller
      operationId: updateExpUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/UserExp"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      tags:
        - Curriculum Vitae
      description: Eliminar experiencia usuario (ID de experiencia)
      # used as the method name of the controller
      operationId: deleteExpUser
      parameters:
        - name: authorization
          in: header
          description: Json Web Token
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
  /token/:id/:
    post:
      tags:
        - Token
      description: Generar JWT - ID de usuario o empresa (1 día)
      # used as the method name of the controller
      operationId: generateToken
      parameters:
        - name: data
          in: body
          required: true
          schema:
            $ref: "#/definitions/JWT"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/OnJobsResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  User:
    type: object
    properties:
      email:
        description: Correo del usuario
        type: string
      name:
        description: Nombre
        type: string
      lastname:
        description: Apellido
        type: string
      rut:
        description: Rut
        type: number
      birthdate:
        description: Fecha de Nacimiento
        type: string
        format: date
      profession:
        description: Profesión
        type: string
      experience:
        description: Experiencia
        type: string
      region:
        description: Region
        type: number
      city:
        description: Ciudad
        type: number
      description:
        description: Descripción
        type: string
      flag:
        description: Recomienda
        type: boolean
      recomendation:
        description: Recomendación
        type: string
      score:
        description: Puntuación
        type: number
...
  UserExp:
    type: object
    properties:
      companyName:
        description: Nombre empresa
        type: string
      title:
        description: Titulo obtenido
        type: string
      location:
        description: Ubicación de la empresa
        type: string
      date:
        description: Fecha de ingreso (yyyy-mm-dd)
        type: string
        format: date
      dateOut:
        description: Fecha de salida (yyyy-mm-dd / True para Actualidad)
        type: string
        format: date
      description:
        description: Descripción
        type: string
...
  JWT:
    type: object
    properties:
      email:
        description: Email de usuario o empresa
        type: string
...
  OnJobsResponse:
    type: string
  ErrorResponse:
    type: string